Adds exceptions from AccountController.

+++ src/main/java/spineapp/backend/controllers/AccountController.java
@@ -4,6 +4,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 import spineapp.backend.daos.AccountDAO;
+import spineapp.backend.daos.AccountTypeDAO;
+import spineapp.backend.exceptions.EmailTakenException;
+import spineapp.backend.exceptions.EntityNotFoundException;
 import spineapp.backend.models.Account;
 
 import java.util.List;
@@ -13,11 +16,13 @@ import java.util.UUID;
 @RequestMapping(path = "api/account")
 public class AccountController {
     private final AccountDAO accountDAO;
+    private final AccountTypeDAO accountTypeDAO;
     private final PasswordEncoder passwordEncoder;
 
     @Autowired
-    public AccountController(AccountDAO accountDAO, PasswordEncoder passwordEncoder) {
+    public AccountController(AccountDAO accountDAO, AccountTypeDAO accountTypeDAO, PasswordEncoder passwordEncoder) {
         this.accountDAO = accountDAO;
+        this.accountTypeDAO = accountTypeDAO;
         this.passwordEncoder = passwordEncoder;
     }
 
@@ -27,7 +32,15 @@ public class AccountController {
     }
 
     @PostMapping
-    public void registerNewAccount(@RequestBody Account account) {
+    public void registerNewAccount(@RequestBody Account account) throws EmailTakenException, EntityNotFoundException {
+
+        if (accountDAO.findAccountByEmail(account.getEmail()).isPresent()) {
+            throw new EmailTakenException();
+        }
+
+        if (!accountTypeDAO.existsById(account.getType().getId())) {
+            throw new EntityNotFoundException(account.getType().getId());
+        }
 
         String encodedPassword = passwordEncoder.encode(account.getPassword());
         account.setPassword(encodedPassword);
@@ -36,7 +49,11 @@ public class AccountController {
     }
 
     @DeleteMapping(path = "{accountId}")
-    public void deleteAccount(@PathVariable("accountId") UUID id) {
+    public void deleteAccount(@PathVariable("accountId") UUID id) throws EntityNotFoundException {
+        if (accountDAO.existsById(id)) {
+            throw new EntityNotFoundException(id);
+        }
+
         accountDAO.deleteAccount(id);
     }
